==> 2_individual_assay_analyses/dual_ipa/2_analyses/1_create_pdest_layouts.ipynb <==
# Create files mapping FACS well and plate to basic allele metadata

This script takes cherry picking and allele info files as input and produces a table relating FACS well and plate IDs to allele IDs and symbols. 

import pandas as pd
import polars as pl
import numpy as np
import os

dualipa_inputs = "../1_inputs"
dualipa_outputs = "../3_outputs"
allele_collection_inputs = "../../../1_allele_collection/1_inputs"
allele_collection_outputs = "../../../1_allele_collection/3_outputs"

Vidal lab ran experiment once and analyzed the data. Some alleles didn't pass the qa/qc thresholds or got no data. Batch5 is is a re-run for all alleles that failed in the first set of batches. Here, we process everything together from scratch. An allele should always be compared to the wt on the same batch and plate. The summary stats can then be compared across batches. 

# Read in cherry picking files for both the initial batches and the second round
cp_col = ['orf_id', 'mut_id', 'node_type','dest_pla_id', 'dest_pla', 'dest_well']

pdest_layout_initial_batches_df = pd.read_csv(f'{dualipa_inputs}/cherry_picking_files/summary_CP_VUSDUALXP.csv', index_col=0)[cp_col]
pdest_layout_b5_df = pd.read_csv(f'{dualipa_inputs}/cherry_picking_files/summary_CP_VUSDUALXP_batch5.csv', index_col=0)[cp_col]

pdest_layout_df = pd.concat([pdest_layout_b5_df, pdest_layout_initial_batches_df])
pdest_layout_df = pl.DataFrame(pdest_layout_df)
pdest_layout_df = pdest_layout_df.with_columns(
    pl.col("orf_id").cast(pl.Int64).alias("orf_id"),
    pl.col("mut_id").cast(pl.Int64).alias("mut_id")
)
pdest_layout_df
### This doesn't have well A01

seq_confirm_code_df = pd.read_table(f"{allele_collection_inputs}/sequence_confirmation_class_code.tsv")
with pd.option_context('display.max_columns', None):
    display(seq_confirm_code_df)

seq_confirm_res_df = pd.read_table(f"{allele_collection_outputs}/VarChampSeqConfirmationResult_Alleles_Pass_QC.tsv")
seq_confirm_res_df = seq_confirm_res_df.rename(
    {
        "orf_id_wt": "orf_id",
        "mutation_id_old": "mut_id"
    }, 
    axis=1
)
with pd.option_context('display.max_columns', None):
    display(seq_confirm_res_df)

seq_confirm_res_df = pl.DataFrame(seq_confirm_res_df)
display(seq_confirm_res_df)

pdest_layout_seq_conf_df = pdest_layout_df.join(
    seq_confirm_res_df, 
    on=["orf_id","mut_id"],
    how="left"
)
print(pdest_layout_seq_conf_df["dualip_sequence_confirmation_class"].unique())

pdest_layout_seq_conf_df = pdest_layout_seq_conf_df.with_columns(
    pl.col("aa_change").fill_null("WT"),
    pl.col("nt_change").fill_null("WT")
)

pdest_layout_seq_conf_df.write_csv(f'{dualipa_outputs}/dualipa_experimental_layout.csv')

# Get allele metadata
# allele_info_df = pd.read_table(f"{allele_collection_inputs}/VarChamp_Consolidated_allele_collection.tsv")
# other_cols = [i for i in allele_info_df.columns if "mutation_id_" not in i]
# mut_cols = [i for i in allele_info_df.columns if "mutation_id_" in i]
# allele_info_df = allele_info_df.melt(id_vars=other_cols, 
#                                      value_vars=mut_cols, 
#                                      var_name='mut_type', 
#                                      value_name='mut_id')
# allele_info_df
# Get gene-level metadata
# gene_info_df = pd.read_csv(f"{allele_collection_outputs}/orf_info.csv")
# gene_info_df = gene_info_df[['orf_id', 'symbol']]
# orf_to_symbol = gene_info_df.set_index("orf_id")["symbol"].to_dict()
# orf_to_symbol = {str(k): v for k, v in orf_to_symbol.items()}

# print(pdest_layout_seq_conf_df["orf_id"].unique())

## Pandas version:
## QC with allele collection df for consistency
## Skipped for future
# meta_cols = ['orf_id', 'mut_id', 'aa_change', 'nt_change', 'symbol']
# # pdest_layout_seq_conf_df = pdest_layout_seq_conf_df.merge(allele_info_df[meta_cols], on=['orf_id', 'mut_id'], how='left', suffixes=("", "_ac"))
# # pdest_layout_seq_conf_df
# # assert(all(pdest_layout_seq_conf_df.dropna(subset="symbol")["symbol"]==pdest_layout_seq_conf_df.dropna(subset="symbol")["symbol_ac"]))
# # assert(all(pdest_layout_seq_conf_df.dropna(subset="aa_change")["aa_change"]==pdest_layout_seq_conf_df.dropna(subset="aa_change")["aa_change_ac"]))
# # assert(all(pdest_layout_seq_conf_df.dropna(subset="nt_change")["nt_change"]==pdest_layout_seq_conf_df.dropna(subset="nt_change")["nt_change_ac"]))
# # pdest_layout_seq_conf_df = pdest_layout_seq_conf_df.drop(columns=["symbol_ac","aa_change_ac","nt_change_ac"])

# ## replace NAs
# pdest_layout_seq_conf_df["orf_id"] = pdest_layout_seq_conf_df["orf_id"].astype("Int64").astype(str)
# pdest_layout_seq_conf_df["orf_id"] = pdest_layout_seq_conf_df["orf_id"].replace("<NA>", np.nan)
# pdest_layout_seq_conf_df

# ## Replace missing symbols
# pdest_layout_seq_conf_df[["aa_change", "nt_change"]] = pdest_layout_seq_conf_df[["aa_change", "nt_change"]].fillna("WT")
# pdest_layout_seq_conf_df["symbol"] = pdest_layout_seq_conf_df["symbol"].fillna(pdest_layout_seq_conf_df["orf_id"].map(orf_to_symbol))
# pdest_layout_seq_conf_df.to_csv(f'{dualipa_outputs}/dualipa_experimental_layout.csv', index=None)
