==> 2_individual_assay_analyses/dual_ipa/2_analyses/2_compile_facs.ipynb <==
# Compile single-cell FACS into minimal dataframe

This script compiles the raw FACS results from each well into one dataframe and writes it out as a compressed parquet file. 

from glob import glob
from os import stat
import pandas as pd
import numpy as np
import polars as pl
import os

dualipa_inputs = "../1_inputs"
dualipa_outputs = "../3_outputs"

n_cell_threshold = 800
wt_gfp_threshold = 100

# Download raw FACS data from Zenodo if haven't done so
# bash_file = f"{dualipa_inputs}/download_raw_FACS.sh"
# os.system(f"bash {bash_file}")

# Compile all data into single df
facs_data = f"{dualipa_inputs}/raw_inputs/FACS_single_cell_data_mCherry_positive_threshold_500/*"
all_df = []
for file_name in glob(facs_data):
    pla_id = file_name.split('/')[-1].split('_')[-3]
    well = file_name.split('/')[-1].split('_')[-2]
    if stat(file_name).st_size > 0:
        df = pd.read_csv(file_name,sep='\t',header=None).rename(columns={0:'GFP',1:'mCherry'})
        df['pla'] = int(pla_id)
        df['well'] = well
        df['n_cells'] = len(df)
        all_df.append(df)

pDEST_DUAL_df = pd.concat(all_df)
# Replace zeros with minimum non-zero value to avoid undefined after taking log
min_nonzero_val = pDEST_DUAL_df[pDEST_DUAL_df["GFP"] > 0]["GFP"].min()
pDEST_DUAL_df.loc[pDEST_DUAL_df["GFP"] == 0, "GFP"] = min_nonzero_val

pDEST_DUAL_df = pl.DataFrame(pDEST_DUAL_df)

# Get GFP:mCherry ratio after replacing zeros with min-value
pDEST_DUAL_df = pDEST_DUAL_df.with_columns(
    (pl.col("GFP")/pl.col("mCherry")).alias("GFP_mCherry_ratio")
)

# Calculate means and medians while grouping by plate and well
agg_df = pDEST_DUAL_df.group_by(["pla", "well"]).agg(
    pl.col("GFP").mean().alias('avg_gfp'),
    pl.col("mCherry").mean().alias("avg_mcherry"),
    pl.col("GFP").median().alias("median_gfp"),
    pl.col("mCherry").median().alias("median_mcherry"),
    pl.col("GFP_mCherry_ratio").mean().alias("avg_GFP_mCherry_ratio"),
    pl.col("GFP_mCherry_ratio").median().alias("median_GFP_mCherry_ratio")
)
pDEST_DUAL_df = pDEST_DUAL_df.join(agg_df, on=["pla", "well"])#.to_pandas()

pDEST_DUAL_df

# Merge with minimal metadata
pdest_layout_df = pl.read_csv(f"{dualipa_outputs}/dualipa_experimental_layout.csv")
pDEST_DUAL_df = pDEST_DUAL_df.join(pdest_layout_df, 
                                   left_on=['pla','well'], 
                                   right_on=['dest_pla_id','dest_well'], 
                                   suffix="_sc"
)
pDEST_DUAL_df = pDEST_DUAL_df.with_columns(
    pl.concat_str([pl.col("pla").cast(pl.String), pl.col("well").cast(pl.String)], separator="_").alias("coordinates"),
    pl.when(
        ((pl.col("n_cells")>=n_cell_threshold) & (pl.col("mut_id")>0))
        | ((pl.col("mut_id") == 0) & (pl.col("avg_gfp") >= wt_gfp_threshold))
    )
    .then(pl.lit(True))
    .otherwise(pl.lit(False))
    .alias("valid_well_no_seq_conf")
)
# pDEST_DUAL_df["coordinates"] = pDEST_DUAL_df["dest_pla_id"].astype(str) + "_" + pDEST_DUAL_df["well"].astype(str)

# pdest_layout_df["dualip_ref_sequence_confirmation_class"].unique()
pDEST_DUAL_df.filter(pl.col("valid_well_no_seq_conf"))

# define valid wells
# Convert columns to integer (NULL -> None)
columns_to_cast = [
    "entry_sequence_confirmation_class",
    "dualip_sequence_confirmation_class",
    "dualip_ref_sequence_confirmation_class",
]

pDEST_DUAL_seq_confirm_df = pDEST_DUAL_df.with_columns([
    pl.col(col).cast(pl.Float64, strict=False) for col in columns_to_cast
])

# Define masks for allele filtering
# At least one entry sequence vector is good
any_expression_good = (
    pl.col("entry_sequence_confirmation_class").is_in([1, 2])
)

mutant_mask = (pl.col('n_cells') >= n_cell_threshold) & (pl.col('mut_id') > 0)
entry_good = pl.col("dualip_sequence_confirmation_class").is_in([1, 2])
# entry_discard = pl.col("dualip_sequence_confirmation_class").is_in([3, 4, 5, 6])
entry_null_like = pl.col("dualip_sequence_confirmation_class").is_in([7, 99]) | pl.col("dualip_sequence_confirmation_class").is_null()
mutant_mask_all = mutant_mask & (entry_good | (entry_null_like & any_expression_good))

wt_mask = (pl.col('mut_id') == 0) & (pl.col('avg_gfp') >= wt_gfp_threshold)
## Define masks for allele filtering not applicable for WT
# entry_ref_good = pl.col("dualip_ref_sequence_confirmation_class").is_in([1, 2])
# entry_ref_null_like = pl.col("dualip_ref_sequence_confirmation_class").is_in([7, 99]) | pl.col("dualip_ref_sequence_confirmation_class").is_null()
# wt_mask_all = wt_mask & (entry_ref_good | (entry_ref_null_like & any_expression_good))

# Combine logic
pDEST_DUAL_seq_confirm_df = pDEST_DUAL_seq_confirm_df.with_columns(
    pl.when(mutant_mask_all | wt_mask)
    .then(True)
    .otherwise(False)
    .alias("valid_well")
)

# pDEST_DUAL_df['valid_well'] = (mutant_mask | wt_mask) & class_mask

pDEST_DUAL_seq_confirm_df.filter(pl.col("valid_well"))

pDEST_DUAL_seq_confirm_df.filter(pl.col("mut_id")==2638)

# write out single-cell data
# pDEST_DUAL_df.to_parquet(f"{dualipa_outputs}/facs_single_cell.parquet", compression="zstd")
pDEST_DUAL_seq_confirm_df.write_parquet(f"{dualipa_outputs}/facs_single_cell.parquet", compression="zstd")
