==> 2_individual_assay_analyses/dual_ipa/2_analyses/3_compute_allele_level_stats.ipynb <==
# Compute abundance change for each allele

This script takes in the single-cell FACS data as input, applies various filtering criteria, and then produces a z-score for the abundance of each allele relative to the reference on the same plate.

import pandas as pd
import numpy as np
from random import choice
import matplotlib.pyplot as plt
import polars as pl
import process_dualipa as proc

dualipa_inputs = "../1_inputs"
dualipa_outputs = "../3_outputs"
meta_outputs = "../../../1_allele_collection/3_outputs"

n_cell_threshold = 800
wt_gfp_threshold = 100

# Read in single-cell measurements
pDEST_DUAL_df = pd.read_parquet(f"{dualipa_outputs}/facs_single_cell.parquet")

# Filter to only keep wells with more than 800 cells
pDEST_DUAL_df = pDEST_DUAL_df[pDEST_DUAL_df['n_cells'] > n_cell_threshold]

# get the mean and median measurements per well
keep_cols = ['symbol', 'node_type', 'nt_change','aa_change', 'pla', 'well', 'coordinates', 'n_cells', 'orf_id', 'mut_id', 'valid_well']
pDEST_DUAL_avg_df = pDEST_DUAL_df[keep_cols + ['avg_gfp', 'avg_mcherry','avg_GFP_mCherry_ratio']].drop_duplicates()
pDEST_DUAL_median_df = pDEST_DUAL_df[keep_cols + ['median_gfp', 'median_mcherry','median_GFP_mCherry_ratio']].drop_duplicates()

## 1. Compute the mean and mediam scores

__Functions for computing is in ```process_dualipa.py```__

Maxime's notes:

    Instead of using Georges' approach to compute the assay's variability, which uses a step with random pairings, 
    Luke suggested to compute a STD from the Log2FC of all individual WT measurements, relative to the mean of the WT of each gene.
    This is a more robust approach, as it does not rely on random pairings.
    This function computes the STD of the Log2FC of all individual WT measurements, relative to the mean of the WT of each gene.
    Returns a tuple with the mean and the STD of the Log2FC of all individual WT measurements, relative to the mean of the WT of each gene.

# _d objects: mean or median GFP_mCherry ratio of each unique (wt orf, plate) combination
# _l objects: list of WT:WT abundances (GFP:mCherry ratios) to estimate assay variability
wt_avg_d, wt_ratio_l = proc.get_wt_variability_d(pDEST_DUAL_avg_df, pDEST_DUAL_df)
wt_median_d, wt_ratio_l_median = proc.get_wt_variability_d_median(pDEST_DUAL_median_df, pDEST_DUAL_df)

wt_std, wt_mean = proc.wt_log2fc_variability(pDEST_DUAL_avg_df, pDEST_DUAL_df)

# Must be two WT replicates on same plate to compute the ratio, otherwise we return a NaN
pDEST_DUAL_avg_allele_df = proc.get_pDEST_DUAL_avg_allele_df(pDEST_DUAL_avg_df, wt_avg_d, wt_ratio_l, wt_std, wt_mean)
pDEST_DUAL_median_allele_df = proc.get_pDEST_DUAL_median_allele_df(pDEST_DUAL_median_df, wt_median_d, wt_ratio_l_median)

mean_df = pl.DataFrame(pDEST_DUAL_avg_allele_df)
median_df = pl.DataFrame(pDEST_DUAL_median_allele_df)

# append metadata and write out files
metadata = pl.read_csv(f"{meta_outputs}/varchamp_alleles_with_pillar_annotations.tsv", separator="\t", infer_schema_length=1000000).with_columns(
    pl.col("orf_id").cast(pl.Int64).alias("orf_id"),
    pl.col("mut_id").cast(pl.Int64).alias("mut_id")
)

median_df.shape

mean_df.shape

## 1.1 Merge the same allele together

median_meta_df = median_df.with_columns(
    pl.col("orf_id").cast(pl.Int64).alias("orf_id"),
    pl.col("mut_id").cast(pl.Int64).alias("mut_id")
).filter(
    pl.col("valid_well"),
    pl.col("zscore_log2").is_not_null()
).join(
    metadata.select(pl.col(["orf_id","mut_id"]+[c for c in metadata.columns if c not in median_df.columns])), on=["orf_id", "mut_id"], how="left"
)
print(median_meta_df.shape)

median_df_dup = median_meta_df.filter(pl.col("gene_variant").is_duplicated())
median_df_no_dup = median_meta_df.filter(~pl.col("gene_variant").is_duplicated())

median_df_dup = median_df_dup.group_by("symbol","nt_change","aa_change","gene_variant").agg(
    [pl.col(col).mean().alias(col) for col in median_df.columns if "median" in col or "zscore_log2" in col]
)
median_df_dup = median_df_dup.join(
    median_meta_df.select(["symbol","nt_change","aa_change","gene_variant"]+[col for col in median_meta_df.columns if col not in median_df_dup and col not in median_df]),
    on=["symbol","nt_change","aa_change","gene_variant"],
    how="left"
).unique(subset=["symbol","nt_change","aa_change","gene_variant"])

median_final_df = pl.concat([
    median_df_no_dup.select(pl.col(median_df_dup.columns)),
    median_df_dup,
], how="vertical_relaxed")

mean_meta_df = mean_df.with_columns(
    pl.col("orf_id").cast(pl.Int64).alias("orf_id"),
    pl.col("mut_id").cast(pl.Int64).alias("mut_id")
).filter(
    pl.col("valid_well"),
    (pl.col("allele_wt_log2fc_zscore").is_not_null()) | (pl.col("zscore_log2").is_not_null())
).join(
    metadata.select(pl.col(["orf_id","mut_id"]+[c for c in metadata.columns if c not in median_df.columns])), on=["orf_id", "mut_id"], how="left"
)
print(mean_meta_df.shape)

mean_df_dup = mean_meta_df.filter(pl.col("gene_variant").is_duplicated())
mean_df_no_dup = mean_meta_df.filter(~pl.col("gene_variant").is_duplicated())

mean_df_dup = mean_df_dup.group_by("symbol","nt_change","aa_change","gene_variant").agg(
    [pl.col(col).mean().alias(col) for col in mean_df.columns if "allele_wt_log2fc_zscore" in col or "zscore_log2" in col]
).rename(
    {"zscore_log2fc"}
)
mean_df_dup = mean_df_dup.join(
    mean_meta_df.select(["symbol","nt_change","aa_change","gene_variant"]+[col for col in mean_meta_df.columns if col not in mean_df_dup and col not in mean_df]),
    on=["symbol","nt_change","aa_change","gene_variant"],
    how="left"
).unique(subset=["symbol","nt_change","aa_change","gene_variant"])

mean_final_df = pl.concat([
    mean_df_no_dup.select(pl.col(mean_df_dup.columns)),
    mean_df_dup,
], how="vertical_relaxed")

mean_final_df.unique("gene_variant")

median_final_df.unique("gene_variant")

median_final_df.write_csv(f'{dualipa_outputs}/DUALIPA_median_zscore.csv')
mean_final_df.write_csv(f'{dualipa_outputs}/DUALIPA_mean_zscore.csv')

# mean_meta_df = mean_df.with_columns(
#     pl.col("orf_id").cast(pl.Int64).alias("orf_id"),
#     pl.col("mut_id").cast(pl.Int64).alias("mut_id")
# ).filter(
#     pl.col("valid_well"),
#     pl.col("zscore_log2").is_not_null()
# ).join(
#     metadata.select(pl.col(["orf_id", "mut_id"]+[c for c in metadata.columns if c not in mean_df.columns])), on=["orf_id", "mut_id"], how="left"
# )
# mean_df.write_csv(f'{dualipa_outputs}/DUALIPA_mean_zscore.csv')
